mem_read(), mem_write() bug fixed; example4 update

+++ examples/firmware_examples/example4-analog-wave-gen/PRU0/main_pru0.c
@@ -5,8 +5,8 @@
 
 #define PRU_SHARED 0x00010000
 
-static void *pruSharedMemory;
-static unsigned int *pruSharedMemory_int;
+static void* pruSharedMemory;
+static uint8_t* pruSharedMemory_int;
 
 extern void start(void);
 
@@ -14,25 +14,26 @@ void main(void){
     // Generate sine values in an array     
     
     int i;
-    unsigned int waveform[100]; 
+    uint8_t waveform[100]; 
     float gain = 50.0f;
     float phase = 0.0f;
     float bias = 50.0f;
     float freq = 2.0f * 3.14159f / 100.0f;
     
     for (i = 0; i < 100; i++){
-        waveform[i] = (unsigned int)(bias + (gain * sin((i*freq) + phase)));
+        waveform[i] = (uint8_t)(bias + (gain * sin((i*freq) + phase)));
+        //waveform[i] = (uint8_t)i;
     }
     
     //uint16_t* sram_pointer = (uint16_t *) PRU_SHARED;
     pruSharedMemory = (void *) PRU_SHARED;
 
-    pruSharedMemory_int = (unsigned int *)pruSharedMemory;
+    pruSharedMemory_int = (uint8_t *)pruSharedMemory;
 
-    unsigned int samplestep = 1;    //delay factor
+    uint8_t samplestep = 1;    //delay factor
     *(pruSharedMemory_int) = samplestep;
 
-    unsigned int numbersamples = 100;
+    uint8_t numbersamples = 100;
     *(pruSharedMemory_int+1) = numbersamples;
 
     for(i = 0; i < 100; i++){
